+++ app/Nova/ResourceRegistrar.php
@@ -29,7 +29,7 @@ use App\Nova\Resources\Main\AnnouncementResource;
 use App\Nova\Resources\Main\Blog;
 use App\Nova\Resources\Main\BlogTag;
 use App\Nova\Resources\Main\Collection;
-use App\Nova\Resources\Main\CollectionItem;
+use App\Nova\Resources\Main\CollectionGroup;
 use App\Nova\Resources\Main\CommentReply;
 use App\Nova\Resources\Main\Comments;
 use App\Nova\Resources\Main\PopupResource;
@@ -79,7 +79,7 @@ class ResourceRegistrar
             Blog::class,
             BlogTag::class,
             Collection::class,
-            CollectionItem::class,
+            CollectionGroup::class,
             Recipe::class,
             RecipeAllergens::class,
             RecipeNutritionalInformation::class,
