+++ app/Nova/Resources/Main/Collection.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace App\Nova\Resources\Main;
 
 use App\Models\Collections\Collection as CollectionModel;
+use App\Models\Collections\CollectionGroup as CollectionGroupModel;
 use App\Nova\Resource;
 use App\Nova\Support\Panels\VisibilityPanel;
 use Illuminate\Database\Eloquent\Builder;
@@ -38,7 +39,7 @@ class Collection extends Resource
 
     public static $search = ['id', 'title'];
 
-    public static $with = ['items', 'media'];
+    public static $with = ['groups', 'groups.items', 'media'];
 
     public function authorizedToView(Request $request)
     {
@@ -132,9 +133,12 @@ class Collection extends Resource
                     ->nullable(),
             ]),
 
-            HasMany::make('Items', 'items', CollectionItem::class),
+            HasMany::make('Groups', 'groups', CollectionGroup::class),
 
-            Text::make('Items', fn ($model) => $model->items->count())
+            Text::make('Groups', fn ($model) => $model->groups->count())
+                ->exceptOnForms(),
+
+            Text::make('Items', fn ($model) => $model->groups->sum(fn(CollectionGroupModel $group) => $group->items->count()))
                 ->exceptOnForms(),
 
             DateTime::make('Created At')->exceptOnForms(),
